requirements.txt
# 核心数据与估值
yfinance>=0.2.28
pandas>=2.0.0
pyyaml>=6.0

# 网页界面
streamlit>=1.28.0

# 交互式图表（估值区间图、FCF预测图、敏感性热力图）
plotly>=5.18.0

# 科学计算（反向DCF用 brentq 精确求根，替代手写二分法）
scipy>=1.11.0
